add usertest

+++ db/migrate/20180402133849_create_users.rb
@@ -5,7 +5,7 @@ class CreateUsers < ActiveRecord::Migration[5.1]
       t.string :password_digest
       t.date :day_of_birth
       t.float :height
-      t.float :weight
+      t.integer :weight
       t.integer :strength
 
       t.timestamps
